Require toolchain deps to appear in toolchain locations
Summary: Tag toolchain rules, and make sure toolchain deps use toolchain rules. Finding a choke point to put the check was a bit hard, but relying on the fact that the second execution platform is/isn't set seemed good. It's maybe a little later in the process than we'd like (ideally it would be attribute resolution, but that isn't possible), but it does capture the notion of ignoring important data, or having important data you don't use, namely the exec_cfg.

Reviewed By: stepancheg

Differential Revision: D37522305

fbshipit-source-id: 16ab466ad8559f1bee1221f91fef97f97bdd155a