buck2: stop sending events from HeartbeatGuard before dropping ServerCommandContext
Summary:
Like it says in the title. This ensures that we stop events at a predictable
time (when we drop the ServerCommandContext), instead of stopping them at the
earliest possible time after the HeartbeatGuard is dropped.

This eliminates a race condition where our behavior is different depending on
whether an event gets fired between dropping ServerCommandContext and dropping
HeartbeatGuard or not.

Also, while in there, I removed the use of an `Abortable` / `AbortHandle` that
wasn't necessary (`JoinHandle` already has a `abort` method).

Reviewed By: ndmitchell

Differential Revision: D37600093

fbshipit-source-id: 6cc75c0bcca5abc312b42a904532880b409e7f6f