buck2: daemon: tie the HeartbeatGuard's lifetime to ServerCommandContext
Summary:
When we drop ServerCommandContext, some state may be lost, notably RE stats.
Once that has happened, it's a good idea to stop sending heartbeat events.

This diff refactors so that the HeartbeatGuard no longer outlives the
ServerCommandContext. The next diff will ensure the HeartbeatGuard no longer
emits once the ServerCommandContext is dropped.

Reviewed By: ndmitchell

Differential Revision: D37600094

fbshipit-source-id: b0ed5845eb61a9451fe9e8a7bcc1a19997adb7b5