Allow building toolchain_dep's directly
Summary:
The is_toolchain_rule checks that a toolchain dependency is only called when someone wants a toolchain dependency. Unfortunately, that means that you can't build one at the root, so can't build and that breaks various automated building systems. For now, disable that check, since I'm not really sure where you could put this to make it work as desired.

I plan to talk to Chris about where such a test should go and reenable it.

Reviewed By: krallin

Differential Revision: D37600899

fbshipit-source-id: 223180e47ea56922e25c2712f72a2a578b77dda0