`rust_python_extension` does no loger automatically add cpython dependency [Part 1 of 2]
Summary:
While checking build of rust-python extensions I've observed that `fbsource//third-party/rust:cpython` dependency is implicitly added into the TARGETS dependencies even if we are building the bindings via pyo3.
In such case having `:cpython` dependency does increase the build time without any real benefit.

In D37375588 I was suggesting to add a flag to control which among `:cpython` and `:pyo3` dependency should be implicitly added but considering that most of the buck targets have explicit `:cpython` or `:pyo3` dependency there is no need to add the implicit dependency (which alters the buck-target definer's expectation).

 ---

This diff is considered **part 1 of 2** because we need to first ensure that all the targets depending on `rust_python_extension` are making explicit definition of the dependencies, the 2nd part will re-iterate the checks in this diff and ultimately remove the implicit addition of dependency.

Reviewed By: zertosh

Differential Revision: D37401415

fbshipit-source-id: 4295c0466b37eac11e69c76390983b0ffe468199