Properly forward passing test details to the buck2 reporter when necessary
Summary:
For some reason that escapes me (maybe to reduce memory consumption of I/O heavy tests?), `from_execution_success` is also in charge to drop the details of passing tests (when the pass_has_details flag is set/unset).

This was always set to false in buck2, which made us ignore details of all passing tests. This diff fixes this forwarding the proper value depending on whether the flag --print-passing-details is set/unset.

Reviewed By: krallin

Differential Revision: D37593536

fbshipit-source-id: 6d557d25fed1d0b5779a1d66457f387e5774e99f