Always print test stdout/err when available
Summary:
Currently, buck2 only prints details of listing failures and tests that are either fails or fatals.

This diff changes the contract so that all test types are printed, and details are always shown when available.

This way we can control from the test runner side what we want to print, e.g., honouring flags like `--print-passing-details` and `--no-details`

Reviewed By: krallin

Differential Revision: D37593535

fbshipit-source-id: 509fcb77fe1085bc0d1ed79e70f7b1a3c5250ebd