Add recursive toolchain_dep resolution
Summary: Change the toolchain_dep to resolve to a list of compatible execution platforms, and cache that on the DICE graph. That is probably faster (avoid reconfiguring the toolchain nodes a lot), and allows us to have toolchains depending on toolchains (a feature we want for Go toolchains at least).

Reviewed By: krallin

Differential Revision: D37605324

fbshipit-source-id: 8812da6e985f998439714e090a385814ce61fb6c