Add a trait that indicates a type's representation in starlark
Summary:
Adding a new trait that will show the string that should be used in type annotations in users' starlark files. This is mostly going to be used for documentation generation.

The reason that this cannot be just part of `StarlarkValue`, is that we want to be able to deduce starlark types of rust types that are only AllocValue/UnpackValue and not StarlarkValue impls themselves.

The blanket implementation for StarlarkValue is presented here along with a few other basic implementations.

Reviewed By: bobyangyf

Differential Revision: D37543944

fbshipit-source-id: 8bcc246067a3caf5b9d6713f89bb139b00bb180d