Extract starlark type reprsentations in docstring generation for `starlark_module`
Summary:
Instead of just doing stringify! as we do now, print the actual starlark type representation in documentation.

To have lifetimes and such for types, this does desugar into a function like this in order to print a type:

```
{
  fn f<#lifetimes_from_type>() {
    #type_goes_here::starlark_type_repr()
  }
  f()
}
```

for each parameter. That's not ideal, but I didn't see a good way to get this with lifetimes otherwise.

A small function was also added so that one can use any named lifetimes they want in the type. e.g. both ValueOf<'v> and ValueOf<'x> would work with this code.

Reviewed By: bobyangyf

Differential Revision: D37543947

fbshipit-source-id: a7001f18a0c71939c7977f0d9c6f684f4bfd4da5