Migrate some java stuff to tsets
Summary:
This migrates the main classpath-related things in java libraries to
use tsets and tset projections.

This saves 47s on analysis of messenger (from 97s total for `audit providers`,
probably ~21s of that is loading, so really like a 2.6x speedup).

It also cuts memory use from 21.3GB to 17.7GB for analysis (since it has args
projections, it likely also has a similar or larger cut to dice storage
requirements for action deps+rdeps after execution).

Reviewed By: mykola-semko

Differential Revision: D37125706

fbshipit-source-id: 25fe0b5bb1e1c94250e7f0a78289ae47a09a5275