Make query attrs operate on ConfiguredTargetNode
Summary:
Now that ConfiguredTargetNode has and exposes directly its deps nodes
(instead of just labels), we can now implement the query attributes query
environment in terms of those nodes instead of needing to build up the parallel
ConfiguredGraphNodes (that query environment required this because it needs to
be able to access the deps of nodes without adding dice edges to them).

With that, we can delete the entire ConfiguredGraphNode bits. We can't quite
have the query attributes use the CqueryEnvironment as there are still ways in
which they are different.

Reviewed By: stepancheg

Differential Revision: D37571926

fbshipit-source-id: 118e8e813faf44256f4366a16a0edf55e0a03af6