CLI-side event aggregator that bundles scuba-relevant events for Scribe
Summary:
**TL;DR Some Scuba columns have been tough to add because the data is only available on the CLI, _or_ because ingress would simultaneously need data from CommandStart, CommandEnd, and instant events emitted during the build. This diff adds an CLI-side event subscriber that maintains state, then sends a single large message to Scribe with all context.**

More context:

`buck2_builds` is populated from the contents of the CommandEnd event fired at the end of each build. Currently, all data needed for `buck2_builds` should be included in that CommandEnd event (if the tailer maintains state, that state is liable to be lost in a reboot).

This diff adds a new event type, written directly from the CLI to Scribe. The intention is for ingress to use this new event type, rather than CommandEnd, to populate `buck2_builds`. This event contains the CommandStart event, CommandEnd event, and anything else we'd like to listen for along the way (e.g. the RE session ID). Because it's written from the client, it has access to args, and (not in this diff, but upcoming) should be able to detect the exit code.

fbcode devX remarks that our Scuba tables are missing useful data, such as:
1. The CLI args used to invoke buck2
2. The RE session ID used by the build
3. The exit code of buck2

Unfortunately, if we send all events from the server, each of these are tricky for different reasons:
1. The CLI args aren't known by the daemon. We could send them, and plumb them through to where we send each CommandEnd event, but that seems not-ideal - we're sending args to the daemon, and plumbing them through the daemon, just so we can log them when the command ends.
2. The RE session ID is currently sent to the client in an Instant event, emitted some time after CommandStart. We could plumb this up through the daemon, also sounds a bit messy.
3. Finally, the exit code of buck2 is basically unknowable from the daemon - we could try to infer it (e.g. we could have the client notify the daemon right before it would crash), but mostly this is tough.

None of requires us to be able to send events from the buck2-client to the event-log (although we may want that in the future).

**Note: This diff doesn't write out the exit-code yet - I wanted to get feedback on this approach before I dive deeper.**

Reviewed By: krallin

Differential Revision: D37293872

fbshipit-source-id: 925ab012775453d2c9a9e5e3e58a343a84d57321