Add a tool for getting the primary and secondary classes of a .dex file
Summary:
This tool accepts a list of primary dex patterns, and a list of class names, and outputs a file containing a list of primary dex class names (class names that match one of the primary dex patterns) and a list of secondary dex class names (the remaining classes).

This is needed in the next diff, separated it out here for simplicity.

Reviewed By: mykola-semko

Differential Revision: D37612128

fbshipit-source-id: 9f2521508ab59045fe43b12290f14baea7e37b29