Remove struct.to_json
Summary: Having to_json on the struct is a bit grim, as it means that when you ask `dir(struct(foo: 1))` it returns `["foo","to_json"]` instead of the more sensible `["foo"]`. I appreciate Bazel does this too (and includes `to_proto`), but we have a top-level `json` function, so should do it more sensible.

Reviewed By: bobyangyf

Differential Revision: D36072988

fbshipit-source-id: bb3d4a1e0335564129f6cbb09f5552d04826c526