Add apple_resource_bundle() rules to apple_bundle()
Summary:
# Context

When building iOS apps on Linux, our Linux-based toolchain still requires native Xcode tools for `ibtool`, `actool` and `momc`.

Because Apple toolchains are single-platform (as all the tools have the same platform constraints - e.g., Clang, etc are Linux binaries), that means that any target/rule will be confined to a single matching exec platform (nb: we exclude fat platforms as we do not support those). So, to call into `ibtool`, `actool` and `momc` from a Linux toolchain, we use a `re_macdo` which can run on Linux and just *connects* to Mac RE to execute the action.

# Problem

The problem is that if we use `re_macdo` on Linux RE (because `re_macdo` is part of the Linux-hosted toolchains which only runs on Linux), then things break as `re_macdo` gets stuck as RE has network isolation - `re_macdo` cannot reach the Mac RE servers (allowing RE access from within RE is also problematic as it removes all caching mechanisms).

A secondary problem is that it's just inefficient to be executing a Linux macdo binary that just connects to Mac RE rather than have Buck2 send the underlying actions (e.g., `ibtool`) to Mac RE directly (more code to maintain + less efficient).

# Solution

The solution is to split the resource compilation part of `apple_bundle()` into a `apple_resource_bundle()` and then to always have a macOS-based exec platform selected for `apple_resource_bundle()` by using a macOS-based toolchain.

This still allows `apple_bundle()` to have a Linux toolchain while `apple_resource_bundle()` can have a macOS toolchain.

`apple_resource_bundle()` gets a copy of the relevant fields from `apple_bundle()`, so that all of the code continues working as expected.

**Execution Groups**

Note that if we had Execution Groups, we wouldn't need to do this split. But we need to remove the dependence on `re_macdo` as that's a blocker for efficient Linux builds and paves the way to more straightforward Linux builds.

Reviewed By: chatura-atapattu

Differential Revision: D37602269

fbshipit-source-id: e60d8de908915a1732dd74fbc21db619743deafa