Pointer is a pointer to AValueOrForward, not AValueHeader
Summary:
During GC, `Value` points to `AValueForward`, not to `AValueHeader`.

So change `Pointer::unpack` functions to return `AValueOrForward` instead of `AValueHeader`.

Practically it provides better debug assertions (in `unpack_header_unchecked` function): when we convert `AValueOrForward` to `AValueHeader`, it is debug assertion that the value is not forward. Previously we did cast without assertion, so even in debug mode when used incorrectly it resulted in crash.

Reviewed By: ndmitchell

Differential Revision: D37706520

fbshipit-source-id: 404c33c958439acb19d7f39ecd963c154f6ff1ca