Add `compile_resources_locally` field to `apple_toolchain()`
Summary:
# Context

We want a field `compile_resources_locally` on `apple_toolchain()` that controls whether `ibtool`, `actool` and `momc` are invoked locally (e.g., MacDo requires this and we cannot do it via constraints as then it forces the whole toolchain to run locally -> so we have no caching - see D37683286 for more).

The field should live on `apple_toolchain()` as that's where we define the `ibtool` etc tools, so the inherent knowledge whether they should run locally is there.

# Changes

We just make it possible to pass `compile_resources_locally` to `apple_toolchain()`.

Reviewed By: IanChilds

Differential Revision: D37750637

fbshipit-source-id: 8d7b4337e982c28048772c43eb7e336ccb51853c