Parallelise primary dex class name comparison
Summary:
In Android, we have the concept of "pre-dexing". What this means is that for every `.jar` that we want to put into the APK, we "pre-dex" it - run `d8` on it to create an intermediate output that is then used to create the final DEX. The final DEX-ing step is much faster if it is given `.dex` files rather than `.jar` files, which is why we do it.

Android also has the concept of "primary" and "secondary" dexes. Each APK can tell `buck` which classes it wants to go into the primary DEX by setting some `primary_dex_patterns` - any class that matches the patterns goes into the primary DEX, otherwise it goes into 1 of N secondary DEX files.

Since the `primary_dex_patterns` can vary for each APK, we need to determine at the APK level which classes go where. We do this using dynamic actions - we wait until every intermediate DEX has been created (along with a list of its class names), and then we read the list of class names for each DEX and assign it to the primary DEX or one of the secondary DEX files. However, this assigning is really slow (2-3 minutes for Messenger) - mostly due to `regex_match` AFAICT.

Anyway, this can be sped up if we do the primary dex matching as soon as the intermediate DEX is available (and in parallel), which is what this change does. Then when we actually need to assign a final DEX, we can do that pretty fast.

I'm open to any ideas of better ways to do this.

Reviewed By: cjhopman

Differential Revision: D35432374

fbshipit-source-id: 387892f41ac59cdb225393ab1d2e699cfa1a663f