Do not print "Waiting for Daemon" in LSP
Summary:
Normally, if we don't receive any events for a while, and we haven't started any spans, we want to let the user know that we're still waiting to hear from the server.
However, when doing bidirection communication, we might keep the connection open indefinitely, and also not really have any communication happening. In those cases, the message is not terribly helpful.

So, add a general "config" struct for SimpleConsole with a field saying whether to print this message, and print that message in all cases except when the "lsp" command is run.

Reviewed By: bobyangyf

Differential Revision: D37658796

fbshipit-source-id: c916845b8343ab81ffe5354f9624281d90b35dae