Remove potential deadlock logic in daemon launch logic for Buck2 cli
Summary:
This diff removes a workaround solution that we've added for Windows to avoid deadlocking on Buck2 daemon launch. The gist of the issue is that when we launch a command that can write to a buffer (running `buck2 daemon`), if the buffer fills up it will wait for the buffer to have space to continue. If we do not simultaneously have some mechanism to clear out the buffer, we end up with a deadlock.

We currently only have a workaround method (redirect all stdout/stderr to null) for Windows as this was easily observable due to WIndow's small default buffer size. But as this is a general potential issue and not a Windows specific one, this diff provides a solution for all cases.

The approach is simple: we launch 3 actions (run `buck2 daemon`, read stdout, and read stderr) and join them together.

Reviewed By: krallin

Differential Revision: D37685801

fbshipit-source-id: 429fcae14202612ad14b2347ea3b0c2761c24481