Convert prefer_local to local_only
Summary:
# Context

- Previously, `local_only` parameter would not cause errors if an action was sent to RE (contrary to what its name might suggest)
- After D37551328 (https://github.com/facebookincubator/buck2/commit/ac33480f73f05a6333461cb28cf3c270137fbf80), we introduced `prefer_local` to have the exact semantics as `local_only`
  - And renamed all `local_only` -> `prefer_local` (i.e., note no behavioural change) (D37645428 (https://github.com/facebookincubator/buck2/commit/d89a5b0ad97795ed2c1ac2a3f449db7ff13ac661), D37645429 (https://github.com/facebookincubator/buck2/commit/9df106f4673c2194c1d809eb2e450ccd846e8376), D37644142)
-  Changed `local_only` semantics (D37521393 (https://github.com/facebookincubator/buck2/commit/f85a3a54033ccde209669306fe68c8069b6a04ea)), so that sending a `local_only` action to RE is a build error

# Changes

In this diff, we convert `prefer_local` to `local_only` (i.e., making the guarantees stronger). That's because we assume the original intent was that it would produce a build error.

NOTE: Each change is done in a separate diff, so that we can use CI signal to understand whether the change is safe. It's also easier to revert a change in the rules than block version bump. See D37521383 (https://github.com/facebookincubator/buck2/commit/57952411d70dcb87383371635aa797f4bdfead6f) for more about the rationale.

Reviewed By: d16r

Differential Revision: D37787338

fbshipit-source-id: da0152ce296a645bacf57a5a509905c23cbf3910