Dedupe library targets earlier than before
Summary: Per title. In future we either want to add in the config so that they are all different, or merge them, but that depends upon what happens with the work that princy is doing in buck1 (we just want to match buck1), so holding off on that for now. Instead, just pick one and ignore any targets with the same config (this is what we were doing before but only once we came to writing out the libraries, rather than when we first collected them. This distinction makes a difference in the next diff when we merge modules, including merging targets that are the same but have a different config).

Reviewed By: cute-jumper

Differential Revision: D37851418

fbshipit-source-id: 304af20c7a044ac67e27af9377633fa82018e79b