Add fileset() operator for grouping files in buck2 queries
Summary: Currently, buck2 does not have a way to pass a list of files (not targets) as input to some query operator. In this diff, we add the fileset(...) operator which groups a list of files. For example, ```buck2 query "noop(fileset(a b c))"``` passes the FileSet consisting of the files a, b, and c to the query operator noop() as its first argument. The motivation for fileset() is to implement a new rbuildfiles() command (currently no-op) which requires file lists as arguments.

Reviewed By: cjhopman

Differential Revision: D37797562

fbshipit-source-id: 31e1494c6af1023f553126e77c5b48bca0ffe7bb