Set a host compatible Python host interpreter for cross-builds
Summary:
When performing fbcode cross-builds of Python, make sure we select a `host_interpreter`
that can run on the exec platform while still being compatible with the target platform.
This is done via using a select generated by `get_exec_fbcode_platform()` and evaluated
in the context of the exec platform, to read a target-compatible fbcode platform's
interpreter config value.

Differential Revision: D37897719

fbshipit-source-id: b3f5bdaf3f66af6e2862e5517b95292857ea3bec