buck2: don't implicitly read files and dirs through symlinks
Summary:
Like it says in the title. Right now, our logic for reading PathMetadata is:

- We see if it's an ExternalSymlink. For this, we are careful to check if there
  are intermediate symlinkscomponents on the path.
- If that's not the case, we just try to read.

But, that means that if something is a symlink to inside the repo, we try to
read it as it if were a regular file. This has a few problems:

- We don't track that we read those things (we track that we read the path in
  the symlink, but e.g. Watchman will never give us an update for that because
  this path  does not actually exist on the FS).
- For Eden I/O, this breaks stuff in the following scenario (in
  `path_artifact_value)`:
  - You list a directory `foo` containing a symlink `bar` to a directory `baz`
  - The listing tells you `bar` exists.
  - You try to get the artifact value for `bar`, it says it's a directory
    (because we ignore symlinks)
  - You try to list `bar`, that fails, because actually it was a symlink, not a
    directory.

With this diff, what will happen is that listing `bar` will return a
redirection telling you to list `baz` instead, which *is* a directory.

At a high level, the implementation of this basically boils down  to
splitting `ExternalSymlink::from_disk` into 2 and reusing the first half
(looking for the first symlink) on all codepaths.

Reviewed By: ndmitchell

Differential Revision: D37886934

fbshipit-source-id: 1e948ccd60222b9f8b0f1e41c4cc88f488da1331