Add package (project relative) path to structured data sent to tpx
Summary:
## Context
As part of [Test Hub](https://fb.workplace.com/groups/testinfra.eng/permalink/1772458422924575/), we want to use code hub projects for defining a test suite. What that implies is that for a folder path in a repo (eg. /fbcode/ctp/multisect/ in fbsource), we want to be able to list all the tests under that folder. In order to do that, **we would need to know which file or folder does a test case live in. We have decided to let test runners report that to us together with other test info.**

We have made this work for JestE2E (very easily), and we are keen to also make at least some subset of fbcode work. **What we need from tests built with buck is some folder path (relative to repo root) that is the parent of the file with the test case lives in.** For example, if we have a rust test in `fbcode/ctp/multisect/src/file1.rs`, then we hope to be able to report that the test lives under any of (`fbcode/ctp/multisect/src/file1.rs`, `fbcode/ctp/multisect/src/`, `fbcode/ctp/multisect/` etc...) folder/file. The more specific, the better. Conveniently, the folder path of TARGETS files is definitely the parent folder of all test cases living somewhere under that package, so I think it would be nice to be able to make buck/tpx report that. I think this granularity of being able to decide where tests live at the buck package level is good enough for people to define test suites in a meaningful way.

## This diff

This diff resolves the project relative path of the target using CellResolver from DiceComputations, and pass the folder path string to tpx.

Reviewed By: krallin

Differential Revision: D37857557

fbshipit-source-id: 5b0014671878c80ed8cba6f5605fdc80f2495fef