Remove cmd_args param to args_projections
Summary:
The cmd_args argument used to be required because there was no way for
the function to create a cmd_args itself (as it was only available as
ctx.actions.cmd_args). Now, the function can create it itself.

In addition, we allow the function to return anything that is
cmd_args-appendable (so anything that cmd_args.add()) would accept. This allows us to
avoid the cmd_args object entirely in most projections which gives a small (but
noticeable) memory improvement.

Reviewed By: bobyangyf

Differential Revision: D37666640

fbshipit-source-id: dff8658659c4f2f7d039afe6b6b4a7c4f500e29c