Allow multiple outputs to `--out -` (stdout)
Summary:
When using buck2's `--out` flag, it currently enforces that you are building a single target with a single default output, so for example `buck2 build :a :b --out foo` doesn't make sense. Does this mean write `a`'s output to `foo`, then overwrite it with `b`'s output to `foo`? Does it mean `foo` is supposed to be implicitly created as a directory?

However for the special case of outputting to stdout, I think it is easy to assign sensible and useful behavior. We'll just write the default outputs out to stdout, one after the other, for as many targets as you were building.

This is desirable for scripts that involve building multiple outputs which are meant to be human-readable (like the `…[diag.txt]` and `…[clippy.txt]` outputs we have on Rust rules) or machine-readable in a format that is sensibly concatenated (like `…[diag.json]` and `…[clippy.json]` which are newline-delimited JSON). The `arc rust-check` / `arc rust-clippy` family of scripts would have used this, prior to being rewritten to bxl anyway.

As of {D36580736 (https://github.com/facebookincubator/buck2/commit/174016244704b630f989e6d7f780ccda6f07dec7)} `-` is already a special case meaning `@/dev/stdin` so it makes sense to me that `--out -` would be a special case for `--out /dev/stdout`, as opposed to the current behavior of outputting a file called `-` in the current directory.

**Before:** gotta combine `--show-full-output` + `cut` + `xargs` + `cat` to print multiple outputs:

https://www.internalfb.com/code/fbsource/[bf91bef4dd02b1bc14b98736612fce3c0d2c6949]/fbcode/common/rust/tools/arc_rust/check?lines=92-95

**After:** just `--out -`

```
buck2 build --console=Super - --out=-
```

Reviewed By: krallin

Differential Revision: D37938454

fbshipit-source-id: b777e4c5b99d63fa0ca4c10a1d74f575269fe660