Add custom_attrs to Doc
Summary: Add a field "custom_attrs" to Doc. These are arbitrary fields, and will be set in a future revision by the StarlarkDocs derive macro. This is primarily useful for extra directives for documentation tooling (e.g. what directory structure you want an object's generated docs to be put into)

Reviewed By: ndmitchell

Differential Revision: D37898096

fbshipit-source-id: 46c7950dcaf7a95eb151175b2c7926a6fd5025ef