Workaround to unblock cross-builds using proc macros on v2
Summary:
Rust's proc macros are Rust libraries which are built into DSOs which are
dlopen'd into the Rust compiler process.  As such, these need to be compiled
targeting a platform which is "compatible" with the Rust compiler.

This currently doesn't happen, as proc macro libraries inherit the target
platform.  In native build cases, which are quite common, this generally just
works, but breaks when cross-compiling (e.g. when targeting aarch64 from x86
hosts, the Rust compiler is x86 but the proc macros are built to run on
aarch64).

In v1, this is handled by D23491443, which adds a `rustc_plugin_platform`
to the Rust platform config, so that users can choose a platform to build
proc macros with (and which is "compatible" with the compiler).  As I
understand it, there isn't a great way to support this in v2 and so this diff
adds a workaround based on `configured_alias` and `native.host_info()`
to unblock cross-builds for aarch64.

#forcetdhashing

Reviewed By: psaab, jsgf

Differential Revision: D37909590

fbshipit-source-id: e06994ecd957ba0c30ed0635ca1760f1753fba54