cli: Add a -A (--output-all-attributes) option to query commands
Summary:
Add a new flag `--output-all-attributes` (`-A`) to query commands. This behaves in the same manner as buck1's `-A` flag.

If `-A`/`--output-all-attributes` is specified and `--output-attribute` is not specified, '.*' is used.
If no output attribute flag is specified, just the target name is returned. This is the current behavior.

Additionally, I've added a new artifact group concept to ensure only one of `--output-all-attributes` / `--output-attribute` / `--output-attributes` is specified. If a user tries to specify more than one of them, they will see an error that looks like the following.
```
error: The argument '--output-all-attributes' cannot be used with '--output-attribute <ATTRIBUTE>'

USAGE:
    buck2 uquery --output-all-attributes <QUERY>

For more information try --help
```

Reviewed By: ndmitchell, krallin

Differential Revision: D37971187

fbshipit-source-id: 496b05d9efdd014c7a3e1801a24783d39c10c1c1