Remove an optimisation for selecting execution platform
Summary: Previously if there were toolchain constraints, building on top of toolchain constraints, we took the subset and then subsetted it further. That's nice and efficient, but this is unlikely to be anywhere that matters (we have a handful of such places) and it makes tracking the reasons for toolchain incompatibilities way harder.

Reviewed By: krallin

Differential Revision: D38033315

fbshipit-source-id: 012c60b16e0853d6042ecdd1c98363909962b50e