Make trie remove capable of removing a subtrie
Summary:
Previously, if you had "a/b/c" in DataTree and then asked to remove "a/b", it would still keep "a/b/c". Instead, how we remove "a/b/c" is that when we later insert an entry for "a/b", we automatically remove "a/b/c". This won't work once we have disk backing for materializer state because that's not a trie.
This diff fixes it so that DataTree's remove now removes a subtree when requested. It also returns the removed subtree, not just the leaf value.

Reviewed By: krallin

Differential Revision: D37471958

fbshipit-source-id: 936ad204f0b99ec19e939337b4427f69c4284943