Bump version when an entry gets redeclared after materialization fails
Summary: When an artifact normally gets declared, we update its version so that an old materialization doesn't update the state of the new artifact entry. When an artifact fails to materialize, we currently transition the artifact's entry from Materializing to Declared but don't update the artifact version. I don't think this is currently a bug on the deferred materializer, but adding disk state will make deferred materializer more complex. If we always bump version on declare, we are less likely to get things wrong.

Reviewed By: krallin

Differential Revision: D37401541

fbshipit-source-id: 35012cacca4ecef4ea242436027d51e6ff25b08b