add with_paths to FileTreeIterator and FileTreeIntoIterator
Summary: FileTree always uses path as key but DataTree's iter and into_iter return VecDeque as key. Instead, this diff adds a `with_paths` method that converts the key from a VecDeque of FileNameBuf to a ProjectRelativePathBuf. This with_paths is on a `WithPathIterator` trait, which is implemented for both FileTree's iterators returned by `iter()` and `into_iter()`.

Reviewed By: krallin

Differential Revision: D37535895

fbshipit-source-id: ad010964a691f1846c6ad522b7b173347d6f9101