Make the toolchain_deps constraints a Vec
Summary: Previously we smashed all the toolchain dep dependencies into a single Map. That meant checking if a toolchain dep was valid was better complexity. But given there are probably a couple of execution platforms in most cases, we're doing a lot of extra complexity to save O(n^2) when n=2. Also makes providing good error messages and explanations harder.

Reviewed By: krallin

Differential Revision: D38033314

fbshipit-source-id: 910a4edac540e8cf9a974d599f6f95547c409c4a