Flip the sense of ToolchainConstraints
Summary:
Previously we listed all the execution platforms we were compatible with. Since we know the set of execution platforms is fixed throughout the build, we can flip that on its head to be those we are incompatible with.

That becomes important if we want to explain _why_ we are incompatible later on.

Reviewed By: krallin

Differential Revision: D38033311

fbshipit-source-id: ffe8f4f26bf48337aea8a493857ae4d2a0cc9551