Add a way to convert ExecutionPlatformIncompatibleReason to IncompatiblePlatformReason
Summary: You can think of these things as being a linked list of reasons, where EPIR is the halfway through point. That means if you add in a ConfiguredTargetLabel, you can get back to an IPR. Perhaps the types should be rejigged to make this relationship more explicit, but for now, this is enough for what I want to do.

Reviewed By: krallin

Differential Revision: D38033312

fbshipit-source-id: 2707c387102059172cd1a39b274faf5ca8085cca