Refactor buck2 docs' sidebars
Summary:
This refactors how our sidebars work for buck2 docs to allow us to have files globbed (and thus automatically show up in the sidebar), as well as specify explicit ordering.

We do this by building finding the explicitly listed files in each category/dir, and then doing processing on the globs, removing anything not in the given dir, and not in the list of explicitly listed files.

- Explicitly listed files have their order maintained as written
- The autogenerated glob at the bottom of each category adds any non-explicit files automatically per docusaurus' normal rules
- Directories that should not be in the sidebar are listed in universallyExcludedDirs

During this work, I also moved things underneathe their logical directories. (e.g. things under "Developers" are in the "developers" dir now)

Reviewed By: krallin

Differential Revision: D37971818

fbshipit-source-id: d6bcaf6121cb9976be6b324db7339d4c7133a842