buck2: make `--local-only` work
Summary:
This comes up often. Let's make it do something. This flag will force us to
run things locally rather than throw an error.

Also, there is a bit of a behavior change. If a command sets `prefer_local`,
we'd normally run it locally *only* right now. With this change, we'll run it
locally first but we may fallback to RE if that fails (which is, of course,
fairly unlikely).

I also renamed the methods on `ExecutorPreference` because frankly their names
were not very informative. The thing that matters (`prefers` / `requires`) is
now the first thing in the function name rather than being last (<nothing> /
`only`).

Reviewed By: ndmitchell

Differential Revision: D38070971

fbshipit-source-id: aa701edee5f8978abf24c7306ca8ad9237cdf3fa