Warn when there's concurrent commands
Summary:
This seems to happen more and more.
Warn about it to users in a way that this gets printed to all commands.

We should also get this into our datasets/logview for all commands, but that mechanism doesn't really exist.
But, we should at least show it in console, which helps a bit more than having nothing.

Next steps is to get buck to better handle these and lock like v1, in which case we don't need the logging and provides a proper experience for users

Reviewed By: krallin

Differential Revision: D38134669

fbshipit-source-id: ff2194b9428f8e11a7c5b635bc14391b6dadf4dd