Make clean_output_paths wait on all pending futures
Summary: Currently, cleaning output paths wait on any pending future to finish only if the removed path is a leaf on the tree. If the removed path is a subtree, then we don't wait on any future at all. This diff fixes this by waiting on all futures on the entire subtree.

Reviewed By: krallin

Differential Revision: D37535896

fbshipit-source-id: 340a066f30915befae02345e08ed3f660831b74d