Refactor ProcessingFuture to an enum with values Cleaning and Materializing
Summary:
Currently `ProcessingFuture` on `ArtifactMaterializationData` is just a generic future and doesn't differentiate between futures that do cleaning or materializing.
This changes `ProcessingFuture` to an enum of either `ProcessingFuture` or a `MaterializingFuture` so there is better introspection on what the future is actually doing.
Renamed `SharedProcessingError` to `SharedMaterializingError`, which is only used by `MaterializingFuture`. `ProcessingFuture` uses a `SharedError`.
Also renamed `pending_fut` to `processing_fut`.

Reviewed By: krallin

Differential Revision: D38052529

fbshipit-source-id: c14c36ef9af85b795a632268b3c19383d90fb3bf