Make docs work with starlark_complex_value
Summary: Add an extra constraint to the __generated_documentation function to make sure it implements StarlarkValue (via Self). This makes it so we can use StarlarkDocs derive on starlark_complex_value! types now without having to muck with getting a set of constraints for the impl on the *Gen struct.

Reviewed By: ndmitchell

Differential Revision: D38143785

fbshipit-source-id: 27e6eff3bd7241f1f86abfddc1737221e2d0d9de