Adds `[xctoolchain]` subtarget on `apple_bundle()` and metadata in `[xcode-data]`
Summary:
# Context

When using the Pika toolchains with Xcode, we need to provide a custom `xctoolchain` bundle to Xcode, so that it can properly debug builds. The process will consist of:

**Part 1: Expose `[xctoolchain]` subtarget (this diff)**

Equivalent to computing the corresponding toolchain

https://www.internalfb.com/code/fbsource/[87f59da4157220ee90859922588ac9571a606dd8]/fbobjc/Tools/focus_2/focus_steps/setup_toolchain.py?lines=70

**Part 2: Download and Setup the xctoolchain Bundle (follow-up diff)**

In the Buck2 project generator, we will download the `xctoolchain` bundle and create the relevant symlinks. That's equivalent to this code

https://www.internalfb.com/code/fbsource/[87f59da4157220ee90859922588ac9571a606dd8]/fbobjc/Tools/focus_2/focus_steps/setup_toolchain.py?lines=71-72

And we will also update the Xcode project files to use the new toolchain

https://www.internalfb.com/code/fbsource/[98207c56ebd0801df99134c070042c71684435b7]/fbobjc/Tools/focus_2/project_steps/fixup_xcconfigs.py?lines=120-122

# Changes in This DIff

Expose an `[xctoolchain]` subtarget on any `apple_bundle()` that provides access to the `xctoolchain`. The subtarget is always available, even if an `xctoolchain` is not always present (no need for special casing/error handling by dependents).

In addition, expose a `has_xctoolchain` field on `[xcode-data]` so that we know whether we need to download an `xctoolchain` when using a generated Xcode project.

Reviewed By: chatura-atapattu

Differential Revision: D37821659

fbshipit-source-id: aa181ea8026dfb1f147b52f62ba85c466af3758c