Specify dir/file when removing symlinks for Windows
Summary:
This diff fixes a symlink removal issue seen when running the inplace e2e tests with Windows.

## Issue
Running ` buck test mode/win //buck2/tests/e2e:test_build_inplace -- test_dynamic_output`, succeeds the first time but fails when running it the second time. It fails because it's unable to clean up some symlinks here: buck2/cli/src/daemon/build/mod.rs?lines=319

When dealing with symlinks in Windows, it seems you always have to specify whether it is for a directory or a file. We have a specific logic for creating symlinks in Windows: buck2/buck2_core/src/fs/anyhow.rs?lines=62 .

## Approach
There is a Windows extension of the fs type that exists, but I opted to not use it as it is marked as unstable: https://doc.rust-lang.org/std/os/windows/fs/trait.FileTypeExt.html

Instead, we can just use `metadata` to follow the symlink and check the file type.

Reviewed By: krallin

Differential Revision: D38132555

fbshipit-source-id: 3faccdc2aa9fb8955004ef7816213404759fa8ac