require explicit dependencies
Summary:
if a target t depends on a library a and a depends on library b, until now the modules of b have been accessible to t despite b being only a transitive dependency of t.

given the scenario above, this diff arranges for compile time errors to result from trying to use b from t (requiring that b become an explicit dependency of t to use it from there).

Reviewed By: ndmitchell

Differential Revision: D38213812

fbshipit-source-id: 3c824d43eb0b880502aca58d85cbaf6f077183b3