rename get_bound to get_bound_deprecated in preparation for migration to new function
Summary: The stack above contains changes to add a concept of additional inputes (set of ArtifactGroups) to Starlark*Artifact types. To make the migration easier, this diff renames get_bound to get_bound_deprecated. In the upcoming diffs, the usages will be migrated to one of get_bound_artifact or get_bound_additional_inputs (see the next diff on the exact definitions)

Reviewed By: bobyangyf

Differential Revision: D38035667

fbshipit-source-id: fda1bb53563d9a3421fa8754735c3ec22f180d35